requirements.txt
Django>=4.2,<6.0
requests>=2.31.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0